+++ src/main/java/com/guotu/gt/domain/PermissionOperation.java
@@ -5,8 +5,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import javax.validation.constraints.NotNull;
-
 /**
  * 操作信息表
  */
@@ -19,7 +17,7 @@ public class PermissionOperation {
      * 主键
      */
     @ApiModelProperty("操作编码，主键")
-    private Byte code;
+    private int code;
 
     /**
      *系统默认操作类型：0浏览；1新增；2查询；3删除；4编辑；5审核
